+++ support/cas-server-support-x509-core/src/test/java/org/apereo/cas/adaptors/x509/authentication/principal/X509SerialNumberPrincipalResolverTests.java
@@ -50,6 +50,7 @@ public class X509SerialNumberPrincipalResolverTests {
             .activeAttributeRepositoryIdentifiers(CollectionUtils.wrapSet(IPersonAttributeDao.WILDCARD))
             .build();
         resolver = new X509SerialNumberPrincipalResolver(context);
+        resolver.sexX509AttributeExtractor(new DefaultX509AttributeExtractor());
     }
 
     @Test
