+++ support/cas-server-support-x509-core/src/test/java/org/apereo/cas/adaptors/x509/authentication/principal/X509SubjectDNPrincipalResolverTests.java
@@ -52,6 +52,8 @@ public class X509SubjectDNPrincipalResolverTests {
         resolver = new X509SubjectDNPrincipalResolver(context);
         resolverRFC2253 = new X509SubjectDNPrincipalResolver(context);
         resolverRFC2253.setSubjectDnFormat(X500Principal.RFC2253);
+        resolver.setX509AttributeExtractor(new DefaultX509AttributeExtractor());
+        resolverRFC2253.setX509AttributeExtractor(new DefaultX509AttributeExtractor());
     }
 
     @Test
