+++ support/cas-server-support-x509-core/src/test/java/org/apereo/cas/adaptors/x509/authentication/principal/X509SubjectPrincipalResolverTests.java
@@ -106,6 +106,7 @@ public class X509SubjectPrincipalResolverTests {
 
         val resolver = new X509SubjectPrincipalResolver(context);
         resolver.setPrincipalDescriptor(descriptor);
+        resolver.setX509AttributeExtrator(new DefaultX509AttributeExtractor());
         val certificate = (X509Certificate) CertificateFactory.getInstance("X509").generateCertificate(
             new FileInputStream(certPath));
